improve performance by decreasing count and frequency of sql queries https://github.com/HarpyWar/pvpgn/issues/85

+++ src/bnetd/storage_sql2.cpp
@@ -221,7 +221,7 @@ namespace pvpgn
 				result = sql->query_res(query);
 				if (!result) continue;
 
-				if (sql->num_fields != 2) {
+				if ((unsigned int)sql->num_fields != 2) {
 					sql->free_result(result);
 					continue;
 				}
